Added Vailyn Tool in new Category
Added `Vailyn` (https://github.com/VainlyStrain/Vailyn) to a new Offensive category: **LFI & Path Traversal**

Vailyn is a tool to find Path Traversal & LFI vulnerabilities, and to exploit them. It uses a number of evasive payloads & techniques, and covers multiple attack vectors (query, path, cookie, POST data). Also, it provides a crawler for automation, a possibility to give an authentication cookie if needed, and a GUI for easy interaction. To improve performance, it detects the vulnerability in a first phase, and only uses confirmed working payloads in the exploitation phase (instead of using every combination with every file path).

In general, the tool should be production ready, however, only the newer reverse shell module is still in testing phase.

In case it doesn't meet the quality standards, I'd highly value your feedback. I'd be glad to improve it in any way I can :)